+++ script/distro/canvasLab-v0.7.158.js
@@ -2,7 +2,7 @@
 // @brief: 			HTML5 canvas drawing framework 
 // @author: 		Justin D. Byrne 
 // @email: 			justin@byrne-systems.com 
-// @version: 		0.7.151 
+// @version: 		0.7.158 
 // @license: 		GPL-2.0
 
 "use strict";
@@ -1976,7 +1976,7 @@ class canvasLab
         this.canvas = canvas;
     }
 
-        ////    [ CANVAS ]  ////////////////////////////////////
+    ////    [ CANVAS ]  ////////////////////////////////////
 
         /**
          * Set canvas value
@@ -2063,6 +2063,19 @@ class canvasLab
             return this._canvas.font;
         }
 
+    ////    [ APPLICATION ]    /////////////////////////////
+
+        /**
+         * Returns this application
+         * @readOnly
+         * @function
+         * @return          {Application}                               Canvas Lab application
+         */
+        get application ( )
+        {
+            return this.#application;
+        }
+
     ////    DOM    /////////////////////////////////////////
 
         /**
@@ -2104,6 +2117,21 @@ class canvasLab
 
                            this._canvas.canvas.clientHeight / 2                 // Y coordinate
                        );
+        }
+
+        /**
+         * Clears canvas
+         * @private
+         * @function
+         */
+        clearCanvas ( )
+        {
+            let _canvas = document.getElementById ( this.canvas );
+
+
+            if ( _canvas )  // @TODO: identify why this check has to take place periodically !
+
+                this._canvas.clearRect ( 0, 0, _canvas.width, _canvas.height );
         }
 
         /**
@@ -5467,7 +5495,7 @@ class StrokeCollection
 ////    OBJECTS    /////////////////////////////////////////
  
 /**
- * @class           {Object}            Location 				Location object
+ * @class           {Object}            Position 				Position object
  * @property        {Point}             point 					Original X & Y coordinates for an object's position
  * @property        {number}            distance 				Distance from origin to destination
  * @property        {number}            direction 				Direction to move towards destination; in degrees
@@ -5477,7 +5505,7 @@ class StrokeCollection
  * @property        {HTMLCanvasElement} canvas                  2D canvas context
  * @property        {clObject} 			master 					Master object
  */
-class Location
+class Position
 {
 	_origin    = undefined;
 	_distance  = undefined;
@@ -5488,7 +5516,7 @@ class Location
 	_master    = undefined;
 
 	/**
-	 * Create a Location object
+	 * Create a Position object
 	 * @property        {Point}  point                              Original X & Y coordinates for an object's position
 	 * @property        {number} distance                  			Distance from origin to destination
 	 * @property        {number} direction 							Direction to move towards destination; in degrees
@@ -5653,7 +5681,7 @@ class Location
  * @property        {HTMLCanvasElement} canvas                  2D canvas context
  * @property        {Anchor}            anchor                  Anchor properties
  * @property        {Options}           options                 Options for this object
- * @property        {Location}          location                Location properties
+ * @property        {Position}          position                Position properties
  */
 class Circle
 {
@@ -5668,7 +5696,7 @@ class Circle
 
     _anchor   = new Anchor;
     #options  = new Options;
-    #location = new Location;
+    #position = new Position;
 
     /**
      * Create a Circle object
@@ -5738,7 +5766,7 @@ class Circle
 
             this.#options.shadow  = ( shadow.offset.x != undefined && shadow.offset.y != undefined );
 
-            this.#location.master = this;
+            this.#position.master = this;
     }
 
     ////    [ POINT ]   ////////////////////////////////////
@@ -5929,14 +5957,14 @@ class Circle
     ////    [ LOCATION ]    ////////////////////////////////
 
         /**
-         * Get location properties
+         * Get position properties
          * @public
          * @function
-         * @return          {Location}                                  Location properties
+         * @return          {Position}                                  Position properties
          */
-        get location ( )
+        get position ( )
         {
-            return this.#location;
+            return this.#position;
         }
 
     ////    VALIDATION  ////////////////////////////////////
@@ -6093,7 +6121,7 @@ class Circle
         _rotatePoint ( ) { }
 
         /**
-         * Sets anchor's point against this object's point location
+         * Sets anchor's point against this object's point position
          * @private
          * @function
          */
@@ -6309,7 +6337,7 @@ class Circle
  * @property        {HTMLCanvasElement} canvas                  2D canvas context
  * @property        {Anchor}            anchor                  Anchor properties
  * @property        {Options}           options                 Options for this object
- * @property        {Location}          location                Location properties
+ * @property        {Position}          position                Position properties
  * @extends 		{Circle}
  */
 class Ellipse extends Circle
@@ -6353,7 +6381,7 @@ class Ellipse extends Circle
  * @property        {HTMLCanvasElement} canvas                  2D canvas context
  * @property        {ControlPoints}     controlPoints           Control point properties
  * @property        {Options}           options                 Options for this object
- * @property        {Location}          location                Location properties
+ * @property        {Position}          position                Position properties
  */
 class Line
 {
@@ -6368,7 +6396,7 @@ class Line
 
     #controlPoints = new ControlPoints;
     #options       = new Options;
-    #location      = new Location;
+    #position      = new Position;
 
     /**
      * Create a Line object
@@ -6621,14 +6649,14 @@ class Line
     ////    [ LOCATION ]    ////////////////////////////////
 
         /**
-         * Get location properties
+         * Get position properties
          * @public
          * @function
-         * @return          {Location}                                  Location properties
+         * @return          {Position}                                  Position properties
          */
-        get location ( )
+        get position ( )
         {
-            return this.#location;
+            return this.#position;
         }
 
     ////    & EXTEND &  ////////////////////////////////////
@@ -7130,7 +7158,7 @@ class Line
  * @property        {Anchor}            anchor                  Anchor properties
  * @property        {HTMLCanvasElement} canvas                  2D canvas context
  * @property        {Options}           options                 Options for this object
- * @property        {Location}          location                Location properties
+ * @property        {Position}          position                Position properties
  */
 class Rectangle
 {
@@ -7145,7 +7173,7 @@ class Rectangle
     _canvas = undefined;
 
     #options  = new Options;
-    #location = new Location;
+    #position = new Position;
 
     /**
      * Create a Rectangle object
@@ -7219,7 +7247,7 @@ class Rectangle
 
             this.#options.shadow  = ( shadow.offset.x != undefined && shadow.offset.y != undefined );
 
-            this.#location.master = this;
+            this.#position.master = this;
     }
 
     ////    [ POINT ]   ////////////////////////////////////
@@ -7461,9 +7489,9 @@ class Rectangle
 
     ////    [ LOCATION ]    ////////////////////////////////
 
-        get location ( )
+        get position ( )
         {
-            return this.#location;
+            return this.#position;
         }
 
     ////    VALIDATION  ////////////////////////////////////
@@ -7599,7 +7627,7 @@ class Rectangle
         _rotatePoint ( ) { }
 
         /**
-         * Sets anchor's point against this object's point location
+         * Sets anchor's point against this object's point position
          * @private
          * @function
          * @see             {@link UTILITIES.individual.set.anchorPoint}
@@ -7772,7 +7800,7 @@ class Rectangle
  * @property        {Anchor}            anchor 					Anchor properties
  * @property        {HTMLCanvasElement} canvas                  2D canvas context
  * @property        {Options}  			options 				Options for this object
- * @property        {Location} 			location 				Location properties
+ * @property        {Position} 			position 				Position properties
  * @extends 		{Rectangle}
  */
 class RoundedRectangle extends Rectangle
@@ -7789,7 +7817,7 @@ class RoundedRectangle extends Rectangle
  * @property        {Shadow}            shadow                  Shadow properties
  * @property        {HTMLCanvasElement} canvas                  2D canvas context
  * @property        {Options}           options                 Options for this object
- * @property        {Location}          location                Location properties
+ * @property        {Position}          position                Position properties
  */
 class Text extends Font
 {
@@ -7802,7 +7830,7 @@ class Text extends Font
     _canvas = undefined;
 
     #options  = new Options;
-    #location = new Location;
+    #position = new Position;
 
     /**
      * Create a Text object
@@ -8037,14 +8065,14 @@ class Text extends Font
     ////    [ LOCATION ]    ////////////////////////////////
 
         /**
-         * Get location properties
+         * Get position properties
          * @public
          * @function
-         * @return          {Location}                                  Location properties
+         * @return          {Position}                                  Position properties
          */
-        get location ( )
+        get position ( )
         {
-            return this.#location;
+            return this.#position;
         }
 
     ////    * SUPER *   ////////////////////////////////////
@@ -8435,7 +8463,7 @@ class Text extends Font
  * @property        {HTMLCanvasElement} canvas                  2D canvas context
  * @property        {Anchor}            anchor                  Anchor properties
  * @property        {Options}           options                 Options for this object
- * @property        {Location}          location                Location properties
+ * @property        {Position}          position                Position properties
  */
 class cImage
 {
@@ -8457,7 +8485,7 @@ class cImage
 
     _anchor   = new Anchor;
 	#options  = new Options;
-    #location = new Location;
+    #position = new Position;
 
 	/**
      * Create a cImage object
@@ -8757,14 +8785,14 @@ class cImage
     ////    [ LOCATION ]    ////////////////////////////////
 
         /**
-         * Get location properties
+         * Get position properties
          * @public
          * @function
-         * @return          {Location}                                  Location properties
+         * @return          {Position}                                  Position properties
          */
-        get location ( )
+        get position ( )
         {
-            return this.#location;
+            return this.#position;
         }
 
     ////    VALIDATION    //////////////////////////////////
@@ -8876,7 +8904,7 @@ class cImage
         _rotatePoint ( ) { }
 
         /**
-         * Sets anchor's point against this object's point location
+         * Sets anchor's point against this object's point position
          * @private
          * @function
          * @see             {@link UTILITIES.individual.set.anchorPoint}
@@ -11761,7 +11789,8 @@ class Animation
 
     #options =
     {
-        cache: false
+        cache:  false,
+        active: false
     }
 
     #queue = new Queue;
@@ -11784,6 +11813,8 @@ class Animation
         this.timing = timing;
         this.period = period;
         this.change = change;
+
+        this.#options.active = true;
     }
 
     ////    [ OBJECT ]    //////////////////////////////////
@@ -11901,7 +11932,7 @@ class Animation
             return this._change;
         }
 
-    ////    [ CHANGE ]    //////////////////////////////////
+    ////    [ CACHE ]    ///////////////////////////////////
 
         /**
          * Set cache
@@ -11969,25 +12000,25 @@ class Animation
          * @function
          * @param           {clObject} object                           Canvas Lab object
          */
-        _cacheObject ( object )
+        _cacheObject ( )
         {
             let _object = undefined;
 
 
-            switch ( object.constructor.name )
+            switch ( this.object.constructor.name )
             {
                 case 'Circle':
 
-                    _object = new Circle ( object.point );
+                    _object = new Circle ( this.object.point );
 
                     // _object = new Circle (
-                    //               object.point,
-                    //               object.radius,
-                    //               object.angle,
-                    //               object.stroke,
-                    //               object.fill,
-                    //               object.shadow,
-                    //               object.canvas
+                    //               this.object.point,
+                    //               this.object.radius,
+                    //               this.object.angle,
+                    //               this.object.stroke,
+                    //               this.object.fill,
+                    //               this.object.shadow,
+                    //               this.object.canvas
                     //           );
 
                     break;
@@ -11995,13 +12026,13 @@ class Animation
                 case 'Ellipse':
 
                     _object = new Ellipse (
-                                  object.point,
-                                  object.radius,
-                                  object.angle,
-                                  object.stroke,
-                                  object.fill,
-                                  object.shadow,
-                                  object.canvas
+                                  this.object.point,
+                                  this.object.radius,
+                                  this.object.angle,
+                                  this.object.stroke,
+                                  this.object.fill,
+                                  this.object.shadow,
+                                  this.object.canvas
                               );
 
                     break;
@@ -12009,13 +12040,13 @@ class Animation
                 case 'Rectangle':
 
                     _object = new Rectangle (
-                                  object.point,
-                                  object.aspect,
-                                  object.round,
-                                  object.stroke,
-                                  object.fill,
-                                  object.shadow,
-                                  object.canvas
+                                  this.object.point,
+                                  this.object.aspect,
+                                  this.object.round,
+                                  this.object.stroke,
+                                  this.object.fill,
+                                  this.object.shadow,
+                                  this.object.canvas
                               );
 
                     break;
@@ -12023,13 +12054,13 @@ class Animation
                 case 'RoundedRectangle':
 
                     _object = new RoundedRectangle (
-                                  object.point,
-                                  object.aspect,
-                                  object.round,
-                                  object.stroke,
-                                  object.fill,
-                                  object.shadow,
-                                  object.canvas
+                                  this.object.point,
+                                  this.object.aspect,
+                                  this.object.round,
+                                  this.object.stroke,
+                                  this.object.fill,
+                                  this.object.shadow,
+                                  this.object.canvas
                               );
 
                     break;
@@ -12037,17 +12068,17 @@ class Animation
                 case 'Text':
 
                     _object = new Text (
-                                  object.point,
-                                  object.text,
-                                  object.type,
-                                  object.size,
-                                  object.weight,
-                                  object.maxWidth,
-                                  object.offset,
-                                  object.stroke,
-                                  object.fill,
-                                  object.shadow,
-                                  object.canvas
+                                  this.object.point,
+                                  this.object.text,
+                                  this.object.type,
+                                  this.object.size,
+                                  this.object.weight,
+                                  this.object.maxWidth,
+                                  this.object.offset,
+                                  this.object.stroke,
+                                  this.object.fill,
+                                  this.object.shadow,
+                                  this.object.canvas
                               );
 
                     break;
@@ -12055,6 +12086,27 @@ class Animation
 
 
             this.#cache.push ( _object );
+        }
+
+        /**
+         * Caches current object
+         * @private
+         * @function
+         */
+        _cache ( )
+        {
+            if ( this.queue.isSet  &&  ! this.queue.isEnd )
+            {
+                if ( this._change.cache )
+
+                    this._cacheObject ( );
+
+
+                this.animate ( );
+            }
+            else
+
+                console.info ( 'animation complete !' );
         }
 
         /**
@@ -12110,6 +12162,24 @@ class Animation
         }
 
         /**
+         * End animation
+         * @private
+         * @function
+         */
+        _end ( )
+        {
+            let _frame = requestAnimationFrame ( this.animate );
+
+
+            cancelAnimationFrame ( _frame );
+
+            this._clearCanvas ( this.object );
+
+
+            return;
+        }
+
+        /**
          * Returns properties animation properties for execution
          * @private
          * @function
@@ -12149,20 +12219,20 @@ class Animation
             let _angle = ( direction % 360 );
 
 
-                _point.x = this.object.location.origin.x + Math.cos ( _angle * Math.PI / 180 ) * distance;
+                _point.x = this.object.position.origin.x + Math.cos ( _angle * Math.PI / 180 ) * distance;
 
-                _point.y = this.object.location.origin.y + Math.sin ( _angle * Math.PI / 180 ) * distance;
+                _point.y = this.object.position.origin.y + Math.sin ( _angle * Math.PI / 180 ) * distance;
 
 
             return _point;
         }
 
         /**
-         * Set Location data
+         * Set Position data
          * @private
          * @function
          */
-        _setLocationData ( )
+        _setPositionData ( )
         {
             for ( let _type in this.change )
             {
@@ -12173,17 +12243,17 @@ class Animation
                 {
                     case 'point':
 
-                        this.object.location.origin    = this.object.point;
+                        this.object.position.origin    = this.object.point;
 
-                        this.object.location.distance  = _difference;
+                        this.object.position.distance  = _difference;
 
-                        this.object.location.direction = _difference;
+                        this.object.position.direction = _difference;
 
                         break;
 
                     case 'move':
 
-                        this.object.location.origin = this.object.point;
+                        this.object.position.origin = this.object.point;
 
 
                         _difference.degree          = ( this.change.rotatePoint )
@@ -12196,9 +12266,9 @@ class Animation
                         let _point = this._getPointByDegreeNDistance ( _difference.degree, _difference.distance );
 
 
-                        this.object.location.distance  = _point;
+                        this.object.position.distance  = _point;
 
-                        this.object.location.direction = _point;
+                        this.object.position.direction = _point;
 
                         break;
 
@@ -12277,9 +12347,9 @@ class Animation
 
                         object.point =
                         {
-                            x: object.location.origin.x + ( object.location.distance * progress ) * Math.cos ( object.location.direction ),
+                            x: object.position.origin.x + ( object.position.distance * progress ) * Math.cos ( object.position.direction ),
 
-                            y: object.location.origin.y + ( object.location.distance * progress ) * Math.sin ( object.location.direction )
+                            y: object.position.origin.y + ( object.position.distance * progress ) * Math.sin ( object.position.direction )
                         }
 
                         break;
@@ -12347,6 +12417,16 @@ class Animation
             }
         }
 
+        /**
+         * Cancels animation
+         * @readOnly
+         * @function
+         */
+        get cancel ( )
+        {
+            this.#options.active = false;
+        }
+
     ////    ANIMATE    /////////////////////////////////////
 
         /**
@@ -12360,7 +12440,7 @@ class Animation
 
                 this._checkQueue ( );
 
-                this._setLocationData ( );
+                this._setPositionData ( );
 
             ////    PROPERTIES    //////////////////////////
 
@@ -12370,23 +12450,21 @@ class Animation
 
                 let _period = this._period;
 
-                let _change = this._change;
-
-                let _queue  = this.queue;
-
-                let _cache  = this.#cache;
+                let _active = this.#options.active;
 
             ////    FUNCTIONS    ///////////////////////////
 
-                let _callback    = ( )                  => this.animate ( );
+                let _transition   = ( object, progress ) => this._transition ( object, progress );
 
-                let _transition  = ( object, progress ) => this._transition ( object, progress );
+                let _clearCanvas  = ( object )           => this._clearCanvas ( object );
 
-                let _clearCanvas = ( object )           => this._clearCanvas ( object );
+                let _drawCache    = ( )                  => this._drawCache ( );
 
-                let _drawCache   = ( )                  => this._drawCache ( );
+                let _cacheObject  = ( )                  => this._cacheObject ( );
 
-                let _cacheObject = ( object )           => this._cacheObject ( object );
+                let _cache        = ( )                  => this._cache ( );
+
+                let _end          = ( )                  => this._end ( );
 
             ////////////////////////////////////////////////
             ////    ANIMATE    /////////////////////////////
@@ -12407,10 +12485,9 @@ class Animation
                 /* normalize */ _progress     = ( true && _progress < 0 ) ? 0 : _progress;
 
 
-                            _transition ( _object, _progress );
-
-
                             _clearCanvas ( _object );
+
+                            _transition ( _object, _progress );
 
 
                             _drawCache ( );
@@ -12418,24 +12495,13 @@ class Animation
                             _object.draw ( );
 
 
-                            if ( _timeFraction < 1 )
+                            ( _active ) ? ( _timeFraction < 1 )                 // Resolve
 
-                                requestAnimationFrame ( animate );
+                                              ? requestAnimationFrame ( animate )
 
-                            else
+                                              : _cache ( )
 
-                                if ( _queue.isSet  &&  ! _queue.isEnd )
-                                {
-                                    if ( _change.cache )
-
-                                        _cacheObject ( _object );
-
-
-                                    _callback ( );
-                                }
-                                else
-
-                                    console.info ( 'animation complete !' );
+                                        : _end ( );
                         }
                     );
                 }
@@ -12458,6 +12524,8 @@ class Animation
  */
 class Application
 {
+    _animation = undefined;
+
     /**
      * Application configurations & details
      * @type            {Object}
@@ -12478,8 +12546,8 @@ class Application
             Author:    'Justin Don Byrne',
             Created:   'October, 2 2023',
             Library:   'Canvas Lab',
-            Updated:   'Aug, 27 2024',
-            Version:   '0.7.151',
+            Updated:   'Aug, 30 2024',
+            Version:   '0.7.158',
             Copyright: 'Copyright (c) 2023 Justin Don Byrne'
         }
     }
@@ -12530,6 +12598,19 @@ class Application
 
             this._isInDom = VALIDATION.isInDom;
     }
+
+    ////    [ ANIMATION ]    ///////////////////////////////
+
+        /**
+         * Get animation
+         * @readOnly
+         * @function
+         * @return          {Animation}                                 Animation object
+         */
+        get animation ( )
+        {
+            return this._animation;
+        }
 
     ////    [ CANVAS ]  ////////////////////////////////////
 
@@ -12651,17 +12732,17 @@ class Application
         {
             if ( Array.isArray ( transition ) )
             {
-                let _animation       = new Animation;
+                this._animation       = new Animation;
 
-                    _animation.queue = new Queue ( transition );
+                this._animation.queue = new Queue ( transition );
 
-                    _animation.animate ( );
+                this._animation.animate ( );
             }
             else
             {
-                let _animation = new Animation ( transition.object, transition.timing, transition.period, transition.change );
+                this._animation = new Animation ( transition.object, transition.timing, transition.period, transition.change );
 
-                    _animation.animate ( );
+                this._animation.animate ( );
             }
         }
 }